requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Cors==4.0.0
qrcode==7.4.2
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pytz==2023.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
